Actions could be performed if you provided the correct POST to the script.  Now you have to be an admin for it to happen
git-svn-id: file:///home/oestrich/tmp/img/trunk@433 ea04b438-0a61-0410-93d0-9bc465a3bab0

+++ admin/albums.php
@@ -15,11 +15,11 @@ require_once("../load.php");
 
 $lncln = new lncln();
 
-if(isset($_POST['name'])){
+if(isset($_POST['name']) && $lncln->user->permissions['isAdmin'] == 1){
 	$album = $lncln->addAlbum($_POST['name']);
 }
 
-if($_GET['action'] == "delete"){
+if($_GET['action'] == "delete" && $lncln->user->permissions['isAdmin'] == 1){
 	$lncln->deleteAlbum($_GET['album']);
 }
